implementation of a set associative cache.
Implementation of a set associative cache. This program reads in 16 bit
words in hexidecimal, then parses the words to map them to different
indexes in the set associative cache. Also parses the offset of the
word. Then i record how many hits and misses there are. The program
reads in from a text file of 30,000 16 bit hex words that I created